+++ lib/src/common/extensions.dart
@@ -38,10 +38,17 @@ extension LatLngBoundsExtensions on LatLngBounds {
       (z) {
         final zoomLevel = minZoom + z;
 
-        final nwt = crs!.latLngToPoint(northWest, zoomLevel.toDouble()).unscaleBy(tileSizePoint).floor();
+        final nwt = crs!
+            .latLngToPoint(northWest, zoomLevel.toDouble())
+            .unscaleBy(tileSizePoint)
+            .floor();
         final nw = CustomPoint<int>(nwt.x, nwt.y);
 
-        final set = crs.latLngToPoint(southEast, zoomLevel.toDouble()).unscaleBy(tileSizePoint).ceil() - const CustomPoint(1, 1);
+        final set = crs
+                .latLngToPoint(southEast, zoomLevel.toDouble())
+                .unscaleBy(tileSizePoint)
+                .ceil() -
+            const CustomPoint(1, 1);
         final se = CustomPoint<int>(set.x, set.y);
 
         return List.generate(
@@ -73,7 +80,8 @@ extension LatLngBoundsExtensions on LatLngBounds {
     const earthRadiusKm = 6378.137;
     const toRad = 180 / pi;
     var newX = point.latitude + ((dx / earthRadiusKm) * toRad);
-    var newY = point.longitude + (((dy / earthRadiusKm) * toRad) / cos(point.latitude * (1 / toRad)));
+    var newY = point.longitude +
+        (((dy / earthRadiusKm) * toRad) / cos(point.latitude * (1 / toRad)));
     return LatLng(newX, newY);
   }
 }
@@ -92,7 +100,8 @@ extension LatLngExtensions on LatLng {
     var dLon = (longitude - other.longitude) * toRad;
     var sLat = sin(dLat / 2);
     var sLon = sin(dLon / 2);
-    var a = sLat * sLat + sLon * sLon * cos(other.latitude * toRad) * cos(latitude * toRad);
+    var a = sLat * sLat +
+        sLon * sLon * cos(other.latitude * toRad) * cos(latitude * toRad);
     var c = 2 * asin(sqrt(a));
     return c * earthRadius;
   }